+++ Samples/GitLinks/GitHubLib/Links/AuthorizationsLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/authorizations")]
+    public class AuthorizationsLink : Link { }
+}