+++ Samples/GitLinks/GitHubLib/Links/AvatarLink.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using Newtonsoft.Json.Linq;
+using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/avatar")]
+    public class AvatarLink : Link { }
+
+
+   
+}