+++ Samples/GitLinks/GitHubLib/Links/CurrentUserLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/current_user")]
+    public class CurrentUserLink : Link { }
+}