+++ Samples/GitLinks/GitHubLib/Links/EmojisLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/emojis")]
+    public class EmojisLink : Link { }
+}