+++ Samples/GitLinks/GitHubLib/Links/EventsLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/events")]
+    public class EventsLink : Link { }
+}