+++ Samples/GitLinks/GitHubLib/Links/FeedsLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/feeds")]
+    public class FeedsLink : Link { }
+}