+++ Samples/GitLinks/GitHubLib/Links/FollowingLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/following")]
+    public class FollowingLink : Link { }
+}