+++ Samples/GitLinks/GitHubLib/Links/GistsLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/gists")]
+    public class GistsLink : Link { }
+}