+++ Samples/GitLinks/GitHubLib/Links/HomeLink.cs
@@ -1,0 +1,7 @@
+﻿using Tavis;
+
+namespace GitHubLib
+{
+    [LinkRelationType("http://api.github.com/rels/home")]
+    public class HomeLink : Link { }
+}