+++ src/Link/Link.cs
@@ -22,7 +22,7 @@ namespace Tavis
     /// </summary>
     /// 
     /// 
-    public class Link : LinkAttributes, IRequestFactory, IResponseHandler
+    public class Link : LinkAttributes, IRequestFactory
     {
         // Client Defined Attributes
         public HttpContent Content { get; set; }
@@ -32,7 +32,7 @@ namespace Tavis
         private Dictionary<string, Hint> _Hints = new Dictionary<string, Hint>();
         private UriTemplate _template;
         private DelegatingRequestBuilder _httpRequestBuilder;
-        private DelegatingResponseHandler _httpResponseHandler;
+
 
         public UriTemplate Template
         {
@@ -71,7 +71,6 @@ namespace Tavis
             var newLink = (Link)Activator.CreateInstance(type);
 
             newLink._httpRequestBuilder = _httpRequestBuilder;  // Can these be copied by reference, or does it need to be by-value
-            newLink._httpResponseHandler = _httpResponseHandler;
             newLink._Hints = _Hints;
             newLink.Method = Method;
             newLink.Content = Content;
@@ -97,17 +96,6 @@ namespace Tavis
             }           
         }
         
-        public Task<HttpResponseMessage> HandleResponseAsync(string linkRelation, HttpResponseMessage responseMessage)
-        {
-            if (_httpResponseHandler != null)
-            {
-                return _httpResponseHandler.HandleResponseAsync(this.Relation, responseMessage);
-            }
-            var tcs = new TaskCompletionSource<HttpResponseMessage>();
-            tcs.SetResult(responseMessage);
-            return tcs.Task;
-        }
-    
         public void AddHint(Hint hint)
         {
             _Hints.Add(hint.Name, hint);
@@ -116,38 +104,6 @@ namespace Tavis
         public IEnumerable<Hint> GetHints()
         {
             return _Hints.Values;
-        }
-
-        public void AddResponseHandler(Action<string,HttpResponseMessage> responseHandlerFunc)
-        {
-            AddResponseHandler(new InlineResponseHandler(responseHandlerFunc, _httpResponseHandler));
-        }
-
-        public  void AddResponseHandler(DelegatingResponseHandler responseHandler)
-        {
-
-            if (_httpResponseHandler == null)
-            {
-                _httpResponseHandler = responseHandler;
-            }
-            else
-            {
-                var currentHandler = _httpResponseHandler as DelegatingResponseHandler;
-                if (currentHandler == null) throw new Exception("Cannot add handler unless existing handler is a delegating handler");
-
-                while (currentHandler != null)
-                {
-                    if (currentHandler.InnerResponseHandler == null)
-                    {
-                        currentHandler.InnerResponseHandler = responseHandler;
-                        currentHandler = null;
-                    }
-                    else
-                    {
-                        currentHandler = currentHandler.InnerResponseHandler;
-                    }
-                }
-            }
         }